+++ PSJira/Functions/Get-JiraIssue.ps1
@@ -54,7 +54,19 @@ function Get-JiraIssue
 
         [Parameter(ParameterSetName = 'ByJQL')]
         [Parameter(ParameterSetName = 'ByFilter')]
-        [Int] $MaxResults = 50,
+        [Int] $StartIndex = 0,
+
+        [Parameter(ParameterSetName = 'ByJQL')]
+        [Parameter(ParameterSetName = 'ByFilter')]
+        [Int] $MaxResults = 0,
+
+        # How many issues should be returned per call to JIRA. This parameter
+        # only has effect if $MaxResults is not provided or set to 0. Normally,
+        # you should not need to adjust this parameter, but if the REST calls
+        # take a long time, try playing with different values here.
+        [Parameter(ParameterSetName = 'ByJQL')]
+        [Parameter(ParameterSetName = 'ByFilter')]
+        [Int] $PageSize = 50,
 
         # Credentials to use to connect to Jira
         [Parameter(Mandatory = $false)]
@@ -67,6 +79,17 @@ function Get-JiraIssue
         $server = Get-JiraConfigServer -ConfigFile $ConfigFile -ErrorAction Stop
 
         Write-Debug "[Get-JiraIssue] ParameterSetName=$($PSCmdlet.ParameterSetName)"
+
+        $psName = $PSCmdlet.ParameterSetName
+
+        if (($psName -eq 'ByJQL' -or $psName -eq 'ByFilter') -and $MaxResults -eq 0)
+        {
+            Write-Debug "[Get-JiraIssue] Using loop mode to obtain all results"
+            $MaxResults = 1
+            $loopMode = $true
+        } else {
+            $loopMode = $false
+        }
     }
 
     process
@@ -112,30 +135,49 @@ function Get-JiraIssue
             }
         } elseif ($PSCmdlet.ParameterSetName -eq 'ByJQL') {
 
-            Write-Debug "[Get-JiraMethod] Escaping query and building URL"
+            Write-Debug "[Get-JiraIssue] Escaping query and building URL"
             $escapedQuery = [System.Web.HttpUtility]::UrlPathEncode($Query)
-            $issueURL = "$($server)/rest/api/latest/search?jql=$escapedQuery&validateQuery=true&expand=transitions&maxResults=$MaxResults"
+            $issueURL = "$($server)/rest/api/latest/search?jql=$escapedQuery&validateQuery=true&expand=transitions&startAt=$StartIndex&maxResults=$MaxResults"
 
-            Write-Debug "[Get-JiraMethod] Preparing for blastoff!"
+            Write-Debug "[Get-JiraIssue] Preparing for blastoff!"
             $result = Invoke-JiraMethod -Method Get -URI $issueURL -Credential $Credential
 
             if ($result)
             {
                 # {"startAt":0,"maxResults":50,"total":0,"issues":[]}
 
-                if ($result.total -gt 0)
+                if ($loopMode)
                 {
-                    Write-Debug "[Get-JiraMethod] Converting REST result to Jira issue"
+                    $totalResults = $result.total
+
+                    Write-Debug "[Get-JiraIssue] Paging through all issues (loop mode)"
+                    $allIssues = New-Object -TypeName System.Collections.ArrayList
+
+                    for ($i = 0; $i -lt $totalResults; $i = $i + $PageSize)
+                    {
+                        $percentComplete = ($i / $totalResults) * 100
+                        Write-Progress -Activity 'Get-JiraIssue' -Status "Obtaining issues ($i - $($i + $PageSize))..." -PercentComplete $percentComplete
+                        Write-Debug "[Get-JiraIssue] Obtaining issues $i - $($i + $PageSize)..."
+                        $thisSection = Get-JiraIssue -Query $Query -StartIndex $i -MaxResults $PageSize -Credential $Credential
+                        foreach ($t in $thisSection)
+                        {
+                            [void] $allIssues.Add($t)
+                        }
+                    }
+                    Write-Progress -Activity 'Get-JiraIssue' -Status 'Obtaining issues' -Completed
+                    Write-Output ($allIssues.ToArray())
+                } elseif ($result.total -gt 0) {
+                    Write-Debug "[Get-JiraIssue] Converting REST result to Jira issue"
                     $obj = ConvertTo-JiraIssue -InputObject $result.issues
 
-                    Write-Debug "[Get-JiraMethod] Outputting result"
+                    Write-Debug "[Get-JiraIssue] Outputting result"
                     Write-Output $obj
                 } else {
-                    Write-Debug "[Get-JiraMethod] No results were found for the specified query"
+                    Write-Debug "[Get-JiraIssue] No results were found for the specified query"
                     Write-Verbose "No results were found for the query [$Query]"
                 }
             } else {
-                Write-Debug "[Get-JiraMethod] Invoke-JiraMethod returned no results"
+                Write-Debug "[Get-JiraIssue] Invoke-JiraMethod returned no results"
             }
         } elseif ($PSCmdlet.ParameterSetName -eq 'ByFilter') {
             $filterObj = Get-JiraFilter -InputObject $Filter -Credential $Credential